Divided files in smaller ones to make code more easy to read Added menu class
