+++ create/solid.js
@@ -1,0 +1,7 @@
+import * as THREE from 'https://unpkg.com/three@0.165.0/build/three.module.js';
+import { draw2dPlane } from '../draw/draw2D.js';
+import { draw3dPlane } from '../draw/draw3D.js';
+
+export function createSolid( sideSize, sides, position, plane, length){
+    
+}
